+++ DataAccessTier/Migrations/MiddlewareContextModelSnapshot.cs
@@ -796,8 +796,8 @@ namespace HEAppE.DataAccessTier.Migrations
                     b.Property<long?>("ProjectId")
                         .HasColumnType("bigint");
 
-                    b.Property<long>("ResourceConsumed")
-                        .HasColumnType("bigint");
+                    b.Property<double>("ResourceConsumed")
+                        .HasColumnType("float");
 
                     b.Property<string>("ScheduledJobId")
                         .HasColumnType("nvarchar(max)");
